# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SaveYourMoney API
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000
...
paths:
...
  /recorrencias:
    get:
      tags: [Recurrences]
      summary: Lista recorrências
      responses:
        "200":
          description: Lista de recorrências
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recurrence"
    post:
      tags: [Recurrences]
      summary: Cria recorrência
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceCreate"
      responses:
        "201":
          description: Recorrência criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recurrence"
        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /recorrencias/{id}:
    put:
      tags: [Recurrences]
      summary: Atualiza recorrência
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecurrenceUpdate"
      responses:
        "200":
          description: Recorrência atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recurrence"
        "404":
          description: Recorrência não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Recurrences]
      summary: Exclui recorrência
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Recorrência excluída
        "404":
          description: Recorrência não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /recorrencias/aplicar:
    post:
      tags: [Recurrences]
      summary: Aplica recorrência
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id: { type: integer }
              required: [id]
      responses:
        "200":
          description: Recorrência aplicada
          content:
            application/json:
              schema:
                type: object
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Expense"
                  incomes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Income"
                required: [expenses, incomes]
        "400":
          description: Payload inválido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Recorrência não encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /relatorios/mes:
    get:
      tags: [Reports]
      summary: Relatório mensal
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: Relatório
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
        "400":
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /relatorios/mes/csv:
    get:
      tags: [Reports]
      summary: Relatório mensal em CSV
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: CSV
          content:
            text/csv:
              schema:
                type: string
        "400":
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /relatorios/mes/pdf:
    get:
      tags: [Reports]
      summary: Relatório mensal em PDF
      parameters:
        - in: query
          name: mes
          schema: { type: integer }
        - in: query
          name: ano
          schema: { type: integer }
      responses:
        "200":
          description: PDF
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Expense:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        category_id: { type: integer, nullable: true }
        recurrence_id: { type: integer, nullable: true }
        payment_method: { type: string }
        notes: { type: string, nullable: true }
      required: [id, name, value, month, year]
...
    Income:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        value: { type: number }
        month: { type: integer }
        year: { type: integer }
        confirmed: { type: boolean }
        notes: { type: string, nullable: true }
      required: [id, name, value, month, year, confirmed]
...
    Recurrence:
      type: object
      properties:
        id: { type: integer }
        kind: { type: string }
        name: { type: string }
        value: { type: number }
        start_month: { type: integer }
        start_year: { type: integer }
        interval_months: { type: integer }
        occurrences: { type: integer }
        category_id: { type: integer, nullable: true }
        payment_method: { type: string, nullable: true }
        confirmed: { type: boolean, nullable: true }
        notes: { type: string, nullable: true }
      required: [id, kind, name, value, start_month, start_year, interval_months, occurrences]
    RecurrenceCreate:
      type: object
      properties:
        kind: { type: string }
        name: { type: string }
        value: { type: number }
        start_month: { type: integer }
        start_year: { type: integer }
        interval_months: { type: integer }
        occurrences: { type: integer }
        category_id: { type: integer, nullable: true }
        payment_method: { type: string, nullable: true }
        confirmed: { type: boolean, nullable: true }
        notes: { type: string, nullable: true }
      required: [kind, name, value, start_month, start_year, interval_months, occurrences]
    RecurrenceUpdate:
      type: object
      properties:
        kind: { type: string }
        name: { type: string }
        value: { type: number }
        start_month: { type: integer }
        start_year: { type: integer }
        interval_months: { type: integer }
        occurrences: { type: integer }
        category_id: { type: integer, nullable: true }
        payment_method: { type: string, nullable: true }
        confirmed: { type: boolean, nullable: true }
        notes: { type: string, nullable: true }
      required: [kind, name, value, start_month, start_year, interval_months, occurrences]
...
    MonthlyReport:
      type: object
      properties:
        month: { type: integer }
        year: { type: integer }
        total_expenses: { type: number }
        total_incomes: { type: number }
        balance: { type: number }
        by_category:
          type: object
          additionalProperties:
            type: number
        goals:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              name: { type: string }
              limit_value: { type: number }
              spent: { type: number }
              remaining: { type: number }
            required: [id, name, limit_value, spent, remaining]
      required: [month, year, total_expenses, total_incomes, balance, by_category, goals]
...